backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-chess==1.999
tensorflow==2.16.1
numpy==1.26.4
tflite_runtime==2.14.0
pytest==7.4.0
pytest-asyncio==0.21.1